added pyproject.toml file to install according to DESC CI guidelines
