--- legacy_blinding/blind_2pt_cosmosis/__init__.py
@@ -1,3 +1,0 @@
-# Author: Arthur Loureiro <arthur.loureiro@fysik.su.se> & Jessie Muir <muir@perimeterinstitute.ca>
-
-__version__ = "2023.06.16"
