--- legacy_blinding/blind_2pt_cosmosis/cosmosis_files/default_blinding_template.ini
@@ -1,51 +1,0 @@
-; ############################################################
-; This is a template file that is referenced by blind_2pt_usingcosmosis.py
-;   It is used as a template for running cosmosis to generate
-;   blinding factors to be applied to 2pt data vectors. To adapt
-;   for a specific pipeline:
-;   1. Edit the %include params.ini line to point towards your baseline pipeline file.
-;   2. The values file should be edited to match your fiducial run. Since we want
-;      to blind sigma8, sigma8_input needs to be in the cosmological parameters,
-;      so it may need to be a values file written specifically for blinding. You could
-;      always set it up to %include your standard values file though.
-;      >> the start values for each parameter is what will be used as reference vals
-;         when generating the blinding factors. 
-;   3. Make sure the file in 2PT_FILE points to a file with the format expected by
-;      your pipeline (with the same binning, etc.)
-;   4. The modules list in [pipeline] should match your standard pipeline with a one
-;      addition: you need sigma8_rescale to be able to include sigma8_input as a parameter
-; ############################################################
-
-
-; import baseline parameter file; change this to point towards the params.ini file for
-; whatever main pipeline you want to mimic
-%include ./blind_2pt_cosmosis/cosmosis_files/params.ini
-
-[DEFAULT]
-; this 2PT file is where we'll pull the n(z) info for computing blinding factors
-2PT_FILE = ./blind_2pt_cosmosis/cosmosis_files/sim_fiducial.fits
-;2PT_FILE = pipeline/2pt_sim_1110_baseline_Y3cov.fits 
-;should this be my 2pt measurement DV?
-
-[pipeline]
-; THIS MODULE LIST SHOULD MATCH YOUR SETTINGS FOR YOUR STANDARD ANALYSIS PIPELINE
-quiet=T
-timing=F
-debug=F
-priors = ./blind_2pt_cosmosis/cosmosis_files/priors.ini
-; sigma8_rescale included so we can set sigma8 as input
-modules = consistency camb extrapolate fits_nz lens_photoz_bias fast_pt source_photoz_bias unbiased_galaxies IA pk_to_cl_gg pk_to_cl add_magnification add_intrinsic 2pt_shear 2pt_gal 2pt_gal_shear shear_m_bias 2pt_like
-; if you want to blind sigma_8, set up values file to have sigma8_input as a parameter rather than A_s
-values = ./blind_2pt_cosmosis/cosmosis_files/blinding_values_template.ini
-
-
-[sigma8_rescale]
-file = cosmosis-standard-library/utility/sample_sigma8/sigma8_rescale.py
-
-[test]
-; if the sampler happens to be set to test, make sure it doesn't
-; save any output when we do blinding. We don't want to see the cosmology
-; and 2pt functions that go into making the blinding factor!
-save_dir=
-
-
