+++ legacy_blinding/src/blind_2pt_cosmosis/__init__.py
@@ -1,0 +1,2 @@
+# Author: Arthur Loureiro <arthur.loureiro@fysik.su.se> & Jessie Muir <muir@perimeterinstitute.ca>
+from ._version import __version__
