+++ legacy_blinding/src/blind_2pt_cosmosis/_version.py
@@ -1,0 +1,1 @@
+__version__ = "2023.10.11"
