+++ legacy_blinding/src/blind_2pt_cosmosis/cosmosis_files/blinding_values_template.ini
@@ -1,0 +1,69 @@
+[cosmological_parameters]
+omega_m      =  0.1	0.3	0.9
+h0           =  0.55	0.69	0.91
+omega_b      =  0.03	0.048	0.12
+n_s          =  0.87	0.97	1.07
+sigma_8_input =  0.2  	0.843  	1.6
+A_s          =  0.5e-09	 2.19e-9  5.0e-09
+; omnuh2       =  0.0006	 0.00083 	   0.01 
+omnuh2       =  0.0
+massive_nu   =  0
+massless_nu  =  3.046
+omega_k      =  0.0
+tau          =  0.0697186
+w            =  -1.0
+;Helium mass fraction.  Needed for Planck 
+yhe          = 0.245341
+
+; %include ${METHODSDIR}/cosmosis/de_models/${DEMODEL}.ini
+
+[shear_calibration_parameters]
+m1 = -0.1  0.0  0.1
+m2 = -0.1  0.0  0.1 
+m3 = -0.1  0.0  0.1
+m4 = -0.1  0.0  0.1
+
+[wl_photoz_errors]
+bias_1 = -0.1   0.0   0.1
+bias_2 = -0.1   0.0   0.1
+bias_3 = -0.1   0.0   0.1
+bias_4 = -0.1   0.0   0.1
+
+[lens_photoz_errors]
+bias_1 = -0.05  0.0  0.05
+bias_2 = -0.05  0.0  0.05
+bias_3 = -0.05  0.0  0.05
+bias_4 = -0.05  0.0  0.05
+bias_5 = -0.05  0.0  0.05
+
+[bias_lens]
+b1 = 0.8  1.7   3.0
+b2 = 0.8  1.7   3.0
+b3 = 0.8  1.7   3.0
+b4 = 0.8  2.0    3.0
+b5 = 0.8  2.0    3.0
+
+[mag_alpha_lens]
+; alpha_1 = -4. 0.903125 4.
+; alpha_2 = -4. 0.68572965 4.
+; alpha_3 = -4. 0.65340057 4.
+; alpha_4 = -4. 1.58867862 4.
+; alpha_5 = -4. 1.93754879 4.
+alpha_1 = 0.903125 
+alpha_2 = 0.68572965 
+alpha_3 = 0.65340057 
+alpha_4 = 1.58867862 
+alpha_5 = 1.93754879 
+
+[intrinsic_alignment_parameters]
+z_piv = 0.62
+A1 = -5.0  0.7  5.0
+A2 = -5.0  -1.36  5.0 
+alpha1=-5.0  -1.7  5.0
+alpha2=-5.0  -2.5  5.0
+bias_ta = 0.0  1.0  2.0
+#bias_tt = 1.0
+
+
+
+
