+++ legacy_blinding/src/blind_2pt_cosmosis/cosmosis_files/priors.ini
@@ -1,0 +1,29 @@
+; These lines specify Gaussian priors
+; for the various parameters in the various sections,
+; with the specified mean and std. dev	
+
+[wl_photoz_errors]
+bias_1 = gaussian -0.0037 0.0177
+bias_2 = gaussian -0.0171 0.0150
+bias_3 = gaussian  0.0200 0.0138
+bias_4 = gaussian  0.0224 0.0215
+
+
+[shear_calibration_parameters]
+m1 = gaussian 0.0 0.021
+m2 = gaussian 0.0 0.021
+m3 = gaussian 0.0 0.021
+m4 = gaussian 0.0 0.021
+
+
+[lens_photoz_errors]
+bias_1 = gaussian 0.0 0.01
+bias_2 = gaussian 0.0 0.01
+bias_3 = gaussian 0.0 0.01
+bias_4 = gaussian 0.0 0.01
+bias_5 = gaussian 0.0 0.01
+
+
+[planck]
+a_planck = gaussian 1.0 0.0025
+
